+++ contracts/contract/tests/margin_trading.rs
@@ -1269,3 +1269,132 @@ async fn test_margin_trading_forceclose_dcl() -> Result<()> {
     check!(view burrowland_contract.get_asset(wrap_token_contract.0.id()));
     Ok(())
 }
+
+#[tokio::test]
+async fn test_margin_trading_forceclose_reserve_not_enough() -> Result<()> {
+    let worker = near_workspaces::sandbox().await?;
+    let root = worker.root_account()?;
+
+    let nusdt_token_contract = deploy_mock_ft(&root, "nusdt", 18).await?;
+    let wrap_token_contract = deploy_mock_ft(&root, "wrap", 18).await?;
+    let wrap_reserve_amount = d(20000, 24);
+    check!(wrap_token_contract.ft_mint(&root, &root, wrap_reserve_amount));
+
+    let ref_exchange_contract = deploy_ref_exchange(&root).await?;
+    {
+        check!(nusdt_token_contract.ft_storage_deposit(ref_exchange_contract.0.id()));
+        check!(wrap_token_contract.ft_storage_deposit(ref_exchange_contract.0.id()));
+        check!(ref_exchange_contract.storage_deposit(&root));
+        check!(ref_exchange_contract.extend_whitelisted_tokens(&root, vec![nusdt_token_contract.0.id(), wrap_token_contract.0.id()]));
+    }
+
+    let oracle_contract = deploy_oralce(&root).await?;
+    let burrowland_contract = deploy_burrowland_with_price_oracle(&root).await?;
+    check!(burrowland_contract.storage_deposit(&root));
+    check!(burrowland_contract.add_asset_handler(&root, &wrap_token_contract));
+    check!(burrowland_contract.add_asset_handler(&root, &nusdt_token_contract));
+    check!(wrap_token_contract.ft_storage_deposit(burrowland_contract.0.id()));
+    check!(nusdt_token_contract.ft_storage_deposit(burrowland_contract.0.id()));
+    check!(burrowland_contract.deposit(&wrap_token_contract, &root, d(10000, 24)));
+
+    let alice = create_account(&root, "alice", None).await;
+    check!(ref_exchange_contract.storage_deposit(&alice));
+    check!(burrowland_contract.storage_deposit(&alice));
+
+    assert!(nusdt_token_contract.ft_mint(&root, &alice, d(10000, 6)).await?.is_success());
+    assert!(wrap_token_contract.ft_mint(&root, &alice, d(100000, 24)).await?.is_success());
+
+    check!(ref_exchange_contract.deposit(&nusdt_token_contract, &alice, d(10000, 6)));
+    check!(ref_exchange_contract.deposit(&wrap_token_contract, &alice, d(10000, 24)));
+
+    check!(ref_exchange_contract.add_simple_swap_pool(&root, vec![nusdt_token_contract.0.id(), wrap_token_contract.0.id()], 5));
+    check!(ref_exchange_contract.add_simple_liquidity(&alice, 0, vec![U128(d(10000, 6)), U128(d(1000, 24))], Some(vec![U128(0), U128(0)])));
+
+    check!(view ref_exchange_contract.get_pool(0));
+
+    let supply_amount = d(1000, 18);
+    let extra_decimals_mult = d(1, 12);
+    check!(nusdt_token_contract.ft_mint(&root, &alice, supply_amount * 10));
+    check!(wrap_token_contract.ft_storage_deposit(alice.id()));
+
+    check!(view burrowland_contract.get_margin_account(&alice));
+
+    check!(burrowland_contract.deposit_to_margin(&nusdt_token_contract, &alice, supply_amount / extra_decimals_mult));
+    
+    check!(burrowland_contract.register_margin_dex(&root, ref_exchange_contract.0.id(), 1));
+    check!(burrowland_contract.register_margin_token(&root, nusdt_token_contract.0.id(), 0));
+    check!(burrowland_contract.register_margin_token(&root, wrap_token_contract.0.id(), 1));
+
+    let current_timestamp = worker.view_block().await?.timestamp();
+    check!(logs burrowland_contract.margin_trading_open_position_by_oracle_call(
+        &oracle_contract, price_data(current_timestamp, Some(100000)), &alice,
+        nusdt_token_contract.0.id(), d(1000, 18).into(), wrap_token_contract.0.id(), d(100, 24).into(), nusdt_token_contract.0.id(), d(900, 18).into(),
+        SwapIndication {
+            dex_id: near_sdk::AccountId::new_unchecked(ref_exchange_contract.0.id().to_string()),
+            swap_action_text: serde_json::to_string(&RefV1TokenReceiverMessage::Execute{
+                referral_id: None,
+                client_echo: None,
+                actions: vec![
+                    RefV1Action::Swap(RefV1SwapAction{
+                        pool_id: 0,
+                        token_in: near_sdk::AccountId::new_unchecked(wrap_token_contract.0.id().to_string()),
+                        amount_in: Some(U128(d(100, 24))),
+                        token_out: near_sdk::AccountId::new_unchecked(nusdt_token_contract.0.id().to_string()),
+                        min_amount_out: U128(d(900, 6)),
+                    })
+                ]
+            }).unwrap()
+        }
+    ));
+
+    check!(view burrowland_contract.get_margin_account(&alice));
+    
+    let alice_margin_account = burrowland_contract.get_margin_account(&alice).await?.unwrap();
+    let pos_id = alice_margin_account.margin_positions.keys().collect::<Vec<&String>>()[0].clone();
+    
+    let position_amount = alice_margin_account.margin_positions.get(&pos_id).unwrap().token_p_amount;
+    let min_out_amount = alice_margin_account.margin_positions.get(&pos_id).unwrap().token_d_info.balance + 10u128.pow(20);
+
+    check!(view burrowland_contract.get_asset(nusdt_token_contract.0.id()));
+    check!(view burrowland_contract.get_asset(wrap_token_contract.0.id()));
+
+    check!(logs burrowland_contract.margin_trading_force_close_mtposition_by_oracle_call(
+        &oracle_contract, price_data(current_timestamp, Some(200000)), &root,
+        alice.id(), &pos_id, position_amount, min_out_amount / 2,
+        SwapIndication {
+            dex_id: near_sdk::AccountId::new_unchecked(ref_exchange_contract.0.id().to_string()),
+            swap_action_text: serde_json::to_string(&RefV1TokenReceiverMessage::Execute{
+                referral_id: None,
+                client_echo: None,
+                actions: vec![
+                    RefV1Action::Swap(RefV1SwapAction{
+                        pool_id: 0,
+                        token_in: near_sdk::AccountId::new_unchecked(nusdt_token_contract.0.id().to_string()),
+                        amount_in: Some(U128(position_amount / 10u128.pow(12))),
+                        token_out: near_sdk::AccountId::new_unchecked(wrap_token_contract.0.id().to_string()),
+                        min_amount_out: U128(min_out_amount / 2),
+                    })
+                ]
+            }).unwrap()
+        }
+    ));
+    
+    let alice_account = burrowland_contract.get_margin_account(&alice).await.unwrap().unwrap();
+    assert!(alice_account.margin_positions.is_empty());
+
+    let protocol_debts = burrowland_contract.get_all_protocol_debts().await.unwrap();
+    let wrap_debt = protocol_debts.get(wrap_token_contract.0.id()).unwrap();
+    check!(logs burrowland_contract.deposit_to_reserve(&wrap_token_contract, &root, 100));
+    let wrap_asset = burrowland_contract.get_asset(wrap_token_contract.0.id()).await.unwrap();
+    assert_eq!(wrap_asset.reserved, 0);
+    let protocol_debts = burrowland_contract.list_protocol_debts(vec![wrap_token_contract.0.id()]).await.unwrap();
+    assert_eq!(protocol_debts.get(wrap_token_contract.0.id()).unwrap().unwrap().0, wrap_debt.0 - 100);
+
+    check!(logs burrowland_contract.deposit_to_reserve(&wrap_token_contract, &root, d(10000, 24) - 100));
+    let wrap_asset = burrowland_contract.get_asset(wrap_token_contract.0.id()).await.unwrap();
+    assert_eq!(wrap_asset.reserved, d(10000, 24) - wrap_debt.0);
+    let protocol_debts = burrowland_contract.list_protocol_debts(vec![wrap_token_contract.0.id()]).await.unwrap();
+    assert!(protocol_debts.get(wrap_token_contract.0.id()).unwrap().is_none());
+    assert!(burrowland_contract.get_all_protocol_debts().await.unwrap().is_empty());
+    Ok(())
+}